requirements.txt
pandas>=1.3
scikit-learn>=1.0
joblib>=1.0
matplotlib>=3.3
seaborn>=0.11
shap>=0.41
fastapi>=0.70
uvicorn>=0.17
streamlit>=1.20
plotly>=5.0
scipy>=1.7

# Deep Learning (PyTorch & Transfer Learning)
torch>=2.0
torchvision>=0.15
torch-optim>=0.0.1
Pillow>=9.0

# Testing
pytest>=7.0
pytest-cov>=4.0
pytest-asyncio>=0.21

# Code quality
black>=23.0
isort>=5.0
flake8>=6.0
pylint>=2.0

# Security
bandit>=1.7
safety>=2.0

# Monitoring and logging
python-json-logger>=2.0

# Docker
docker>=6.0

